ios: Parse messages from JSONValue
Switch message and update parsing in SDKBridge to operate directly on
JSONValue trees instead of Any/[String: Any].

JSONValue now supports integer decoding to preserve Int64 semantics
for ids, timestamps, and sequence numbers.

+++ ios/DelightHarnessTests/JSONCodingTests.swift
@@ -4,7 +4,7 @@ import XCTest
 final class JSONCodingTests: XCTestCase {
     func testJSONValueRoundTrip() throws {
         let value: JSONValue = .object([
-            "a": .number(1),
+            "a": .int(1),
             "b": .string("two"),
             "c": .array([.bool(true), .null])
         ])
@@ -22,4 +22,3 @@ final class JSONCodingTests: XCTestCase {
         XCTAssertTrue(pretty?.contains("\n") == true)
     }
 }
-
